+++ src/argilla/client/feedback/integrations/huggingface/model_card/__init__.py
@@ -12,6 +12,28 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from argilla.client.feedback.dataset.local.dataset import FeedbackDataset
+from .model_card import (
+    ArgillaModelCard,
+    FrameworkCardData,
+    OpenAIModelCardData,
+    PeftModelCardData,
+    SentenceTransformerCardData,
+    SetFitModelCardData,
+    SpacyModelCardData,
+    SpacyTransformersModelCardData,
+    TransformersModelCardData,
+    TRLModelCardData,
+)
 
-__all__ = ["FeedbackDataset"]
+__all__ = [
+    "ArgillaModelCard",
+    "FrameworkCardData",
+    "SentenceTransformerCardData",
+    "TransformersModelCardData",
+    "SetFitModelCardData",
+    "PeftModelCardData",
+    "SpacyModelCardData",
+    "SpacyTransformersModelCardData",
+    "OpenAIModelCardData",
+    "TRLModelCardData",
+]